:hammer: Deal with CodeClimate
